dev-requirements.txt
flask-testing
coveralls
pep257
pytest
pytest-cache
pytest-cov
pytest-pep8
httpretty
